Big clean up
Short description

First step for a big clean up.

Why:

Too much dead codes...

This change addresses the need by:

+ deleted *old* equation-abstract-based objects;
+ renamed equation_adv in equation;
+ renamed integrator_rk_implicit in integrator_rk_explicit;
+ general clean up.

Side effects:

Only Flame 1D has been modified to be compilable: other test are likely
not compilable.

Note:

On my laptop with Ubuntu/gfortran 6.2 the test compile with also debug
options... other source of entropy!

Moreover, ifort 16.0.3 that is on the same laptop, compiles the test,
but... does not work (as does ifort 17 on my workstation): it generates
the same identical error of gnu for mesh association, but now it does
for field association!!!!!!